Consumers, Operators & Producers

+++ smartpackaging/src/main/java/pt/ipleiria/estg/dei/ei/dae/smartpackaging/ebjs/UserBean.java
@@ -20,7 +20,7 @@ public class UserBean {
     }
 
     // create user
-    public void create(int user_id, String username, String email, String password, UserRole role) {
+    public void create(Long user_id, String username, String email, String password, UserRole role) {
         User user = new User(user_id, username, email, hashPassword(password), role);
         em.persist(user);
     }
